__main__.py が読みづらかったため、ファイル分割のリファクタリングをした #7
