+++ python/.gitignore
@@ -1,2 +1,4 @@
+__pycache__/
+
 coconala-scraping.txt
 coconala-scraping-div.txt
